+++ src/main/java/org/octri/notification/controller/NotificationController.java
@@ -16,6 +16,9 @@ import org.octri.notification.registry.NotificationStatusRegistry;
 import org.octri.notification.registry.NotificationTypeRegistry;
 import org.octri.notification.repository.NotificationRepository;
 import org.octri.notification.view.NotificationStatusSelectOption;
+import org.octri.notification.view.NotificationViewer;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
@@ -31,6 +34,8 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 @Controller
 @RequestMapping("/admin/notification")
 public class NotificationController extends AbstractEntityController<Notification, NotificationRepository> {
+
+	private static final Logger logger = LoggerFactory.getLogger(NotificationController.class);
 
 	private final NotificationRepository repository;
 	private final IdentifiableEntityFinder<?> recipientFinder;
@@ -62,16 +67,28 @@ public class NotificationController extends AbstractEntityController<Notificatio
 		var template = super.list(model);
 		@SuppressWarnings("unchecked")
 		Iterable<Notification> notifications = (Iterable<Notification>) model.get("entity_list");
+		Map<String, NotificationViewer> viewerByType = notificationTypeRegistry.getRegisteredTypes().stream()
+				.collect(Collectors.toMap(
+						t -> t,
+						t -> notificationTypeRegistry.getHandler(t).getViewer()));
+		// Prepare viewers
+		for (var type : viewerByType.keySet()) {
+			var viewer = viewerByType.get(type);
+			var typeNotifications = StreamSupport.stream(notifications.spliterator(), false)
+					.filter(n -> n.getNotificationType().equals(type))
+					.toList();
+			viewer.prepare(typeNotifications);
+		}
 		List<Notification> notificationViews = StreamSupport
 				.stream(notifications.spliterator(), false)
 				.map(n -> {
-					var handler = notificationTypeRegistry.getHandler(n.getNotificationType());
-					if (handler != null) {
-						n.setNotificationRecipientView(handler.getViewer().getRecipientView(n));
-						n.setNotificationMetadataView(handler.getViewer().getMetadataView(n));
+					var viewer = viewerByType.get(n.getNotificationType());
+					if (viewer != null) {
+						n.setNotificationRecipientView(viewer.getRecipientView(n));
+						n.setNotificationMetadataView(viewer.getMetadataView(n));
 					}
 					return n;
-				}).collect(Collectors.toList());
+				}).toList();
 		model.put("entity_list", notificationViews);
 		model.put("notificationStatuses", notificationStatusRegistry.getStatuses());
 		ViewUtils.addPageScript(model, "table-filtering.js");